requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
beautifulsoup4
python-dotenv
sqlalchemy>=2.0
aiosqlite
yfinance
pandas